+++ include/cxxdbc/concept/statement.h
@@ -1,24 +1,14 @@
 #ifndef INCLUDED_CXXDBC_CONCEPT_STATEMENT_H
 #define INCLUDED_CXXDBC_CONCEPT_STATEMENT_H
 
+// CxxDBC top-level includes
+#include <cxxdbc/statement.h>
+
 // third-party includes
 #include <boost/utility/string_ref.hpp>
 
 namespace cxxdbc {
 namespace concept {
-
-/// Defines the v-table for dynamic dispatch to a driver implementation of the statement concept.
-class IStatement {
-public:  // RAII
-  virtual ~IStatement() = default;
-
-public:  // methods
-  virtual void close() = 0;
-
-  virtual bool execute(::boost::string_ref query) = 0;
-
-  virtual bool isClosed() const = 0;
-};
 
 /// Wraps a driver implementation of the statement concept (Tconcept) and provides lifecycle
 /// management.
@@ -27,7 +17,7 @@ public:  // methods
 template <typename Tconcept>
 class StatementWrapper : public IStatement {
 private:  // variables
-  Tconcept m_concept;
+  Tconcept& m_concept;
 
 public:  // RAII
   StatementWrapper() = default;
@@ -36,12 +26,13 @@ public:  // RAII
 
   StatementWrapper(StatementWrapper &&) = delete;
 
-  explicit StatementWrapper(Tconcept &&m_concept)
-      : m_concept(::std::forward<Tconcept>(m_concept)) {}
+  explicit StatementWrapper(Tconcept &&concept)
+      : m_concept(::std::forward<Tconcept>(concept)) {}
 
   ~StatementWrapper() = default;
 
 public:  // methods
+
   void close() override final { m_concept.close(); }
 
   bool execute(::boost::string_ref query) override final { return m_concept.execute(query); }
